move enums to separated files
